Fix outpoint size (64 bit vs 32 bit value).

+++ include/bitcoin/system/chain/outpoint.hpp
@@ -40,7 +40,7 @@ public:
 
     static constexpr size_t serialized_size() NOEXCEPT
     {
-        return point::serialized_size() + sizeof(uint32_t);
+        return point::serialized_size() + sizeof(uint64_t);
     }
 
     /// Constructors.
